Make settings not jump too much
